backend/requirements.txt
annotated-types==0.7.0
anyio==4.10.0
bcrypt==4.3.0
charset-normalizer==3.4.3
click==8.2.1
fastapi==0.116.2
h11==0.16.0
idna==3.10
joblib==1.5.2
lxml==6.0.1
numpy==2.3.3
pandas==2.3.2
passlib==1.7.4
pillow==11.3.0
pydantic==2.11.9
pydantic_core==2.33.2
PyJWT==2.10.1
PyMuPDF==1.26.4
python-dateutil==2.9.0.post0
python-docx==1.2.0
python-multipart==0.0.20
pytz==2025.2
reportlab==4.4.3
scikit-learn==1.7.2
scipy==1.16.2
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
threadpoolctl==3.6.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uvicorn==0.35.0
psycopg[binary]>=3.1
alembic>=1.13
python-dotenv>=1.0
